JP-1318: Keyword updates (#4602)
* JP-1318: Keyword updates

* Consolidate guider models into a single module

* Move tcatfile/scatfile to core schema; include necessary code updates

* Fix unit tests

* More fixes to tcatfile and scatfile

* A few more updated keyword values

* Fix blend rules for exposure time keywords
