+++ jwst/datamodels/schemas/container.schema.yaml
@@ -8,11 +8,11 @@ properties:
     type: object
     properties:
       table_name:
-        title: Name of ASN table
+        title: Name of the ASN table
         type: string
-        fits_keyword: ASNTAB
+        fits_keyword: ASNTABLE
       pool_name:
-        title: Name of ASN pool used to generate this table
+        title: Name of the ASN pool
         type: string
         fits_keyword: ASNPOOL
       resample:
