+++ jwst/datamodels/schemas/extract1dimage.schema.yaml
@@ -22,7 +22,7 @@ allOf:
       type: integer
       fits_keyword: SPORDER
       fits_hdu: SCI
-    dispersion_axis:
+    dispersion_direction:
       title: Dispersion direction
       type: integer
       fits_keyword: DISPAXIS
