+++ jwst/datamodels/schemas/image.schema.yaml
@@ -37,22 +37,3 @@ allOf:
       datatype: float32
 - $ref: variance.schema
 - $ref: pathlosscorr.schema
-- type: object
-  properties:
-    meta:
-      type: object
-      properties:
-        tweakreg_catalog:
-            type: object
-            properties:
-              filename:
-                title: Output tweakreg catalog filename
-                type: string
-                fits_keyword: TCATFILE
-        source_catalog:
-            type: object
-            properties:
-              filename:
-                title: Output source catalog filename
-                type: string
-                fits_keyword: SCATFILE
