+++ jwst/datamodels/schemas/multiextract1d.schema.yaml
@@ -26,7 +26,7 @@ allOf:
             type: integer
             fits_keyword: SPORDER
             fits_hdu: SCI
-          dispersion_axis:
+          dispersion_direction:
             title: Dispersion direction
             type: integer
             fits_keyword: DISPAXIS
